--- starknet_py/abi/__init__.py
@@ -1,2 +1,0 @@
-from .model import Abi
-from .parser import AbiParser, AbiParsingError
