+++ starknet_py/abi/v0/schemas_test.py
